+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 0.4.19 (unreleased)
 -------------------
 
-- Added utils method to create content from a dictionary, to get object following criterias,
-  to apply multiple transitions
+- Added methods to create content from a dictionary, to get object following criterias,
+  to apply multiple transitions, to get a RichTextValue object
   [sgeulette]
 - Added default views for Dexterity content and container that display
   fields in a table with widget label and the left and widget value on